Got a button working
So I'm starting ever so slightly to get a handle on this.
I've got a working button and some text that is starting
to take shape. Next I need to add callback functions so
that changing a value does something, then I need to figure
out how to change some colors. But one thing at a time...