The exit button works at least
At this point the exit button takes you from the connection
screen to the main screen. So that works at least. Now what
should happen is that once the connect button gets hit the
button name should change and the connection status should
change. But sadly that doesn't seem to be working just yet.
More work to be done I suppose.